Move `product_search_space` to `samplers` module.
