+++ tests/samplers_tests/test_samplers.py
@@ -201,3 +201,43 @@ def test_sample_relative():
     study.optimize(objective, n_trials=10, catch=())
     for trial in study.trials:
         assert trial.params == {'a': 3.2, 'b': 'baz', 'c': 30, 'd': 30, 'e': 30}
+
+
+def test_product_search_space():
+    # type: () -> None
+
+    study = optuna.create_study()
+
+    # No trial.
+    assert optuna.samplers.product_search_space(study) == {}
+
+    # First trial.
+    study.optimize(lambda t: t.suggest_int('x', 0, 10) + t.suggest_uniform('y', -3, 3), n_trials=1)
+    assert optuna.samplers.product_search_space(study) == {
+        'x': IntUniformDistribution(low=0, high=10),
+        'y': UniformDistribution(low=-3, high=3)
+    }
+
+    # Second trial (only 'y' parameter is suggested in this trial).
+    study.optimize(lambda t: t.suggest_uniform('y', -3, 3), n_trials=1)
+    assert optuna.samplers.product_search_space(study) == {
+        'y': UniformDistribution(low=-3, high=3)
+    }
+
+    # Failed or pruned trials are not considered in the calculation of a product search space.
+    def objective(trial, exception):
+        # type: (optuna.trial.Trial, Exception) -> float
+
+        trial.suggest_uniform('z', 0, 1)
+        raise exception
+
+    study.optimize(lambda t: objective(t, RuntimeError()), n_trials=1)
+    study.optimize(lambda t: objective(t, optuna.structs.TrialPruned()), n_trials=1)
+    assert optuna.samplers.product_search_space(study) == {
+        'y': UniformDistribution(low=-3, high=3)
+    }
+
+    # If two trials have the same name but different distributions,
+    # those are regarded as different trials.
+    study.optimize(lambda t: t.suggest_uniform('y', -1, 1), n_trials=1)
+    assert optuna.samplers.product_search_space(study) == {}
